pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-assistant"
version = "1.0.0"
description = "A polished CLI AI assistant with multimodal capabilities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Assistant", email = "assistant@example.com"}
]
keywords = ["ai", "assistant", "cli", "llm", "multimodal", "voice"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pypdf>=4.0.0",
    "click>=8.1.0",
    "prompt-toolkit>=3.0.0",
    "httpx>=0.27.0",
    "ddgs>=9.0.0",
    "pyaudio>=0.2.14",
    "google-genai>=1.0.0",
    "pyttsx3>=2.90",
    "edge-tts>=6.0.0",
    "speechrecognition>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "pygame>=2.5.0",
    "duckduckgo-search>=5.0.0",
]

[project.optional-dependencies]
voice = [
    "pyttsx3>=2.90",
    "SpeechRecognition>=3.10.0",
    "sounddevice>=0.4.0",
]
voice-hq = [
    "edge-tts>=6.0.0",
    "SpeechRecognition>=3.10.0",
    "sounddevice>=0.4.0",
]
all = [
    "pyttsx3>=2.90",
    "edge-tts>=6.0.0",
    "SpeechRecognition>=3.10.0",
    "sounddevice>=0.4.0",
]

[project.scripts]
ai = "src.main:main"
assistant = "src.main:main"

[project.urls]
Homepage = "https://github.com/example/ai-assistant"
Documentation = "https://github.com/example/ai-assistant#readme"
Repository = "https://github.com/example/ai-assistant"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
